Update the openapi.generator plugin from version 4.0.0-beta to 4.0.2 fix issues around paths containing . characters. REFERENCE:  https://github.com/OpenAPITools/openapi-generator/pull/1895
As part of this up rev, certain unit tests began failing related to String and URI conflicts.  After some testing, found that removing the "format: uri" tag from the yaml file resolved the issue.  I did this in an attempt to keep the method signature String given the implied unit test functionality.
I did not do a full audit and it may be recommended that we confirm this before merging by someone more familiar with the code generation.